# Make `Remove()` return a bool indicating whether present