Make `Remove()` return a bool indicating whether the object was present.

This is useful for the thread-safe-set because the underlying mutex
isn't public, so without this it's impossible to know if the element was
originally present.

This is a tweak on https://github.com/deckarep/golang-set/pull/66.

This won't be a breaking change because today there's no returned value... 
so if there starts being a bool returned, legacy code will simply ignore it.

I'm sure this will have a minor perf hit, so putting this up to play
around with. If it's a super minor perf regression, then probably okay,
otherwise better to split `Remove()` from `RemoveWithResult()` or some
such API so that `Remove()` stays as high perf as possible.