@jeffwidman 2022-03-28
This is significantly slower for the `ThreadUnsafe` version... going from ~105ns to ~160ns.

Interestingly, it is **not** much slower for the `ThreadSafe` version... going from ~170ns to ~180ns.

Given that the primary reason for needing a dedicated method is for the thread safe version (the thread unsafe you can simply wrap in a lock, and then look and then remove yourself), then ideally we'd only make `ThreadSafe` return the bool, and keep thread unsafe w/no bool for perf reasons. But that is an inconsistent interface, which is a no-go.

So a separate one-off method added to the ThreadSafe implementation probably makes the most sense... something more like what @firejump implemented in #66.

Details:
```
~/C/o/golang-set ❯❯❯ go test -bench BenchmarkRemove                                                                                                                                                           
goos: darwin
goarch: amd64
pkg: github.com/deckarep/golang-set/v2
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkRemoveSafe-12      	 8835639	       161.6 ns/op
BenchmarkRemoveUnsafe-12    	12152125	       160.2 ns/op
PASS
ok  	github.com/deckarep/golang-set/v2	9.320s
~/C/o/golang-set ❯❯❯ go test -bench BenchmarkRemove                                                                                                                                                              
goos: darwin
goarch: amd64
pkg: github.com/deckarep/golang-set/v2
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkRemoveSafe-12      	 8985301	       183.0 ns/op
BenchmarkRemoveUnsafe-12    	12733628	       164.3 ns/op
PASS
ok  	github.com/deckarep/golang-set/v2	10.086s
~/C/o/golang-set ❯❯❯ go test -bench BenchmarkRemove
goos: darwin
goarch: amd64
pkg: github.com/deckarep/golang-set/v2
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkRemoveSafe-12      	 9495962	       180.3 ns/op
BenchmarkRemoveUnsafe-12    	12763028	       164.3 ns/op
PASS
ok  	github.com/deckarep/golang-set/v2	10.274s
~/C/o/golang-set ❯❯❯ go test -bench BenchmarkRemove
goos: darwin
goarch: amd64
pkg: github.com/deckarep/golang-set/v2
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkRemoveSafe-12      	 9474217	       183.7 ns/op
BenchmarkRemoveUnsafe-12    	12182640	       162.5 ns/op
PASS
ok  	github.com/deckarep/golang-set/v2	10.024s
~/C/o/golang-set ❯❯❯ gco main
Switched to branch 'main'
Your branch is up to date with 'origin/main'.
~/C/o/golang-set ❯❯❯ go test -bench BenchmarkRemove
goos: darwin
goarch: amd64
pkg: github.com/deckarep/golang-set/v2
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkRemoveSafe-12      	10953123	       169.3 ns/op
BenchmarkRemoveUnsafe-12    	18089376	       106.3 ns/op
PASS
ok  	github.com/deckarep/golang-set/v2	12.279s
~/C/o/golang-set ❯❯❯ go test -bench BenchmarkRemove                                                                                                                                                                                    
goos: darwin
goarch: amd64
pkg: github.com/deckarep/golang-set/v2
cpu: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
BenchmarkRemoveSafe-12      	11173486	       171.6 ns/op
BenchmarkRemoveUnsafe-12    	17952781	       104.0 ns/op
PASS
ok  	github.com/deckarep/golang-set/v2	12.210s
```